Add TableManager. Refactor RangerTableManager in terms of TableManager.
